backend/requirements.txt
django>=4.2,<4.3
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=3.13.0
django-filter>=23.2
redis>=5.0
bcrypt>=4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
drf-spectacular>=0.26.4
